refactor: extract supportability period policy
