+++ quality/baseline_report.md
@@ -10,7 +10,7 @@ documentation posture, and validation gates. It is not a completion claim.
 
 ## Code Size Baseline
 
-- Python source files under `src/`: 183
+- Python source files under `src/`: 184
 - Python test files under `tests/`: 93
 - Python packages under `src/`: 9
 - API entry point route/middleware/handler decorators in `src/app/main.py`: 0
@@ -37,8 +37,8 @@ documentation posture, and validation gates. It is not a completion claim.
 | src/app/services/concentration/simulation_resolver.py | 360 | 12246 |
 | tests/unit/test_risk_engine_branch_coverage.py | 353 | 13642 |
 | tests/unit/test_concentration_engine_modes.py | 334 | 12149 |
-| src/app/services/calculation_supportability.py | 328 | 11177 |
 | src/app/services/concentration/parsing.py | 319 | 11074 |
+| tests/integration/test_risk_calculate_live_characterization.py | 314 | 12945 |
 
 ### Largest Functions And Classes
 
@@ -89,12 +89,12 @@ All checks passed!
 - Ruff format check: passed
 
 ```text
-290 files already formatted
+291 files already formatted
 ```
 - Type checking: passed
 
 ```text
-Success: no issues found in 276 source files
+Success: no issues found in 277 source files
 ```
 - Unit coverage snapshot: passed
 
@@ -108,10 +108,10 @@ Success: no issues found in 276 source files
 ...
 src\app\upstream_errors.py                                    55      0     20      3    96%   181->194, 187->189, 192->194
 ------------------------------------------------------------------------------------------------------
-TOTAL                                                       5173     80    978     79    97%
+TOTAL                                                       5178     80    978     79    97%
 
-144 files skipped due to complete coverage.
-474 passed in 8.32s
+145 files skipped due to complete coverage.
+474 passed in 9.97s
 ```
 
 ## Complexity And Maintainability Snapshot
@@ -141,7 +141,7 @@ src\app\services\risk\__init__.py - A (100.00)
 - Dependency hygiene: passed
 
 ```text
-Scanning 183 files...
+Scanning 184 files...
 
 Success! No dependency issues found.
 ```
@@ -240,7 +240,7 @@ tests/unit/test_upstream_errors.py::test_classify_upstream_transport_error_matri
 tests/unit/test_upstream_errors.py::test_invalid_payload_and_missing_data_carry_structured_categories
 tests/unit/test_upstream_errors.py::test_extract_upstream_error_detail_variants
 
-474 tests collected in 1.61s
+474 tests collected in 2.15s
 ```
 - Import-linter report-only: passed
